backend/requirements.txt
Django==2.2.6
djangorestframework==3.10.3
psycopg2-binary==2.8.3
django-health-check==3.11.0
django-environ=0.4.5